[#215] Add options through options loader in kafka api for read, write and fetching offsets
