+++ gimel-dataapi/gimel-connectors/gimel-kafka-2.2/src/main/scala/com/paypal/gimel/kafka2/reader/KafkaStreamConsumer.scala
@@ -1,0 +1,220 @@
+/*
+ * Copyright 2018 PayPal Inc.
+ *
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.paypal.gimel.kafka2.reader
+
+import scala.collection.immutable.Map
+import scala.language.implicitConversions
+
+import org.apache.kafka.common.TopicPartition
+import org.apache.spark.sql._
+import org.apache.spark.sql.functions._
+import org.apache.spark.sql.streaming.DataStreamReader
+import org.apache.spark.streaming.kafka010._
+import spray.json._
+
+import com.paypal.gimel.common.catalog.GimelCatalogJsonProtocol._
+import com.paypal.gimel.datastreamfactory.StructuredStreamingResult
+import com.paypal.gimel.kafka2.conf.{KafkaClientConfiguration, KafkaConstants}
+import com.paypal.gimel.kafka2.utilities.{BrokersAndTopic, KafkaOptionsLoaderUtils, KafkaUtilitiesException, TopicAndProps}
+import com.paypal.gimel.kafka2.utilities.ImplicitKafkaConverters._
+import com.paypal.gimel.kafka2.utilities.KafkaUtilities._
+
+/**
+  * Implements Kafka Stream Consumer Logic here
+  */
+object KafkaStreamConsumer {
+
+  val logger = com.paypal.gimel.logger.Logger()
+
+
+  /**
+    *
+    * Function to get the starting offsets for the stream to read from
+    *
+    * @param conf KafkaClientConfiguration object that holds the configuration paremeters
+    * @param kafkaTopic The kafkaTopics list to subscribe to
+    * @return Starting Offsets in a Map[TopicPartition, Long]
+    */
+  private def getStartOffsets(conf: KafkaClientConfiguration, kafkaTopic: String, brokers: String) = {
+    if (conf.kafkaCustomOffsetRange.isEmpty()) {
+      val lastCheckPoint: Option[Array[OffsetRange]] = getLastCheckPointFromZK(conf.zkHostAndPort, conf.zkCheckPoints)
+      val availableOffsetRange: Array[OffsetRange] = BrokersAndTopic(brokers, kafkaTopic).toKafkaOffsetsPerPartition
+      if (lastCheckPoint == None) {
+        logger.info("No CheckPoint Found !")
+        if(conf.kafkaAutoOffsetReset.equals(KafkaConstants.earliestOffset)) {
+          logger.info("Fetching from the beginning")
+          availableOffsetRange.map {
+            x => (new TopicPartition(x.topic, x.partition) -> x.fromOffset)
+          }.toMap
+        } else {
+          logger.info("Fetching from the latest offset")
+          availableOffsetRange.map {
+            x => (new TopicPartition(x.topic, x.partition) -> x.untilOffset)
+          }.toMap
+        }
+      } else {
+        logger.info(s"Found Checkpoint Value --> ${lastCheckPoint.get.mkString("|")}")
+        lastCheckPoint.get.map {
+          x => (new TopicPartition(x.topic, x.partition) -> x.untilOffset)
+        }.toMap
+      }
+    }
+    else {
+      val customOffsetRangesForStream: Array[OffsetRange] = getCustomOffsetRangeForReader(conf.kafkaTopics.split(","), conf.kafkaCustomOffsetRange, KafkaConstants.gimelAuditRunTypeStream)
+      customOffsetRangesForStream.map {
+        x => (new TopicPartition(x.topic, x.partition) -> x.fromOffset)
+      }.toMap
+    }
+  }
+
+  /**
+   *
+   * Function to get the starting offsets for the stream to read from
+   *
+   * @param conf : KafkaClientConfiguration object that holds the configuration paremeters
+   * @param kafkaOptionsMap : Map of topics to options
+   * @return Starting Offsets in a Map[TopicPartition, Long]
+   */
+  private def getStartOffsets(conf: KafkaClientConfiguration, kafkaOptionsMap: Map[String, Map[String, String]]) = {
+    if (conf.kafkaCustomOffsetRange.isEmpty()) {
+      val lastCheckPoint: Option[Array[OffsetRange]] = getLastCheckPointFromZK(conf.zkHostAndPort, conf.zkCheckPoints)
+      val availableOffsetRange: Option[Array[OffsetRange]] = kafkaOptionsMap.map(x =>
+        TopicAndProps(x._1, x._2).toKafkaOffsetsPerPartition).reduceOption(_.union(_))
+      val availableOffsetRanges: Array[OffsetRange] = availableOffsetRange match {
+        case None =>
+          throw new KafkaUtilitiesException("There is an issue ! No Offset Range From Kafka ... Is the topic having any message at all ?")
+        case Some(value) =>
+          value
+      }
+      if (lastCheckPoint == None) {
+        logger.info("No CheckPoint Found !")
+        if(conf.kafkaAutoOffsetReset.equals(KafkaConstants.earliestOffset)) {
+          logger.info("Fetching from the beginning")
+          availableOffsetRanges.map {
+            x => (new TopicPartition(x.topic, x.partition) -> x.fromOffset)
+          }.toMap
+        } else {
+          logger.info("Fetching from the latest offset")
+          availableOffsetRanges.map {
+            x => (new TopicPartition(x.topic, x.partition) -> x.untilOffset)
+          }.toMap
+        }
+      } else {
+        logger.info(s"Found Checkpoint Value --> ${lastCheckPoint.get.mkString("|")}")
+        lastCheckPoint.get.map {
+          x => (new TopicPartition(x.topic, x.partition) -> x.untilOffset)
+        }.toMap
+      }
+    }
+    else {
+      val customOffsetRangesForStream: Array[OffsetRange] = getCustomOffsetRangeForReader(conf.kafkaTopics.split(","), conf.kafkaCustomOffsetRange, KafkaConstants.gimelAuditRunTypeStream)
+      customOffsetRangesForStream.map {
+        x => (new TopicPartition(x.topic, x.partition) -> x.fromOffset)
+      }.toMap
+    }
+  }
+
+
+  /**
+    *
+    * Function to return the last saved checkpoint from zookeeper
+    *
+    * @param conf KafkaClientConfiguration object that holds the configuration paremeters
+    * @return Optional checkpoint Offsets in a Array[OffsetRange]
+    */
+  private def getLastCheckPoint(conf: KafkaClientConfiguration) = {
+    val lastCheckPoint: Option[Array[OffsetRange]] = getLastCheckPointFromZK(conf.zkHostAndPort, conf.zkCheckPoints)
+    lastCheckPoint
+  }
+
+  /**
+    *
+    * Core Function to create a structured stream
+    *
+    * @param sparkSession the spark session passed by the user
+    * @param conf KafkaClientConfiguration object that holds the configuration paremeters
+    * @return StreamingResult in a StructuredStreamingResult Object
+    */
+  def createStructuredStream(sparkSession: SparkSession, conf: KafkaClientConfiguration, props: Map[String, Any]): StructuredStreamingResult = {
+    try {
+      val isStreamParallel = conf.isStreamParallel
+      val streamParallels = conf.streamParallelismFactor
+      logger.debug(
+        s"""
+           |isStreamParallel --> ${isStreamParallel}
+           |streamParallels --> ${streamParallels}
+      """.stripMargin)
+      // Retrieve kafka options from OptionsLoader if specified
+      val kafkaTopicsOptionsMap : Map[String, Map[String, String]] = KafkaOptionsLoaderUtils.getAllKafkaTopicsOptions(conf)
+      logger.info("kafkaTopicsOptionsMap -> " + kafkaTopicsOptionsMap)
+      logger.info(s"Zookeeper Server : ${conf.zkHostAndPort}")
+      logger.info(s"Zookeeper Checkpoint : ${conf.zkCheckPoints}")
+      val startOffsetsForStream: Map[TopicPartition, Long] = getStartOffsets(conf, kafkaTopicsOptionsMap)
+      val startOffsetsStructured = startOffsetsForStream.toList.groupBy(_._1.topic())
+        .mapValues(_.map(x => (x._1.partition().toString, x._2)).toMap)
+      val topics = conf.kafkaTopics
+
+      val eachKafkaTopicToOptionsMap = KafkaOptionsLoaderUtils.getEachKafkaTopicToOptionsMap(kafkaTopicsOptionsMap)
+      val dataStreamReader: DataStreamReader = sparkSession
+        .readStream
+        .format(KafkaConstants.KAFKA_FORMAT)
+        .option(KafkaConstants.KAFKA_SUBSCRIBE, topics)
+        .options(conf.kafkaConsumerProps)
+        .option(KafkaConstants.KAFKA_START_OFFSETS, startOffsetsStructured.toJson.toString())
+        .option(KafkaConstants.STREAM_FAIL_ON_DATA_LOSS, conf.failOnDataLoss)
+        .options(eachKafkaTopicToOptionsMap.getOrElse(topics, Map.empty[String, String]))
+
+      val kafkaDF = if (conf.maxRatePerTrigger.isEmpty) {
+        dataStreamReader.load()
+      } else {
+        dataStreamReader
+          .option(KafkaConstants.KAFKA_MAX_OFFSETS_PER_TRIGGER, conf.maxRatePerTrigger)
+          .load()
+      }
+
+      val kafkaDFFiltered = if (conf.kafkaSourceFieldsList.toLowerCase == KafkaConstants.allKafkaSourceFields) {
+        kafkaDF
+      } else {
+        val kafkaSourceFieldsList = conf.kafkaSourceFieldsList.split(",")
+        kafkaDF.select(kafkaSourceFieldsList.map(name => col(name)): _*)
+      }
+
+      val kafkaDFParallelized = if (isStreamParallel) {
+        kafkaDFFiltered.repartition(streamParallels).toDF(kafkaDFFiltered.columns: _*)
+      } else {
+        kafkaDFFiltered
+      }
+
+      // CheckPointer Function - CheckPoints each window
+      val saveCheckPoint: Unit = inStructuredStreamCheckPoint(sparkSession, conf.zkHostAndPort, conf.zkCheckPoints)
+      // Provide Option to Clear CheckPoint
+      val deleteCheckPoint: (String) => Unit = clearCheckPoint(conf.zkHostAndPort, conf.zkCheckPoints, _: String)
+      // Provide Option to Get DataFrame for a Simple String Message from Kafka Topic
+
+      StructuredStreamingResult(kafkaDFParallelized, saveCheckPoint, deleteCheckPoint)
+    } catch {
+      case ex: Throwable => {
+        ex.printStackTrace()
+        throw ex
+      }
+    }
+  }
+}
+
